backend/pyproject.toml
[project]
name = "health-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "k10i", email = "kubota.daichi405@mail.kyutech.jp" }
]
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.120.4",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "psycopg2-binary>=2.9.9",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
